refactor: Move node_fns and related items into dedicated mod
